requirements.txt
streamlit
langchain-community
langchain
langchain-groq
langgraph
sqlalchemy
python-dotenv
groq